fastapi/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
python-dotenv==1.0.1
openai==1.19.0
requests==2.31.0
sqlalchemy==2.0.29
psycopg2-binary==2.9.9
pillow==10.3.0
Jinja2==3.1.3
boto3==1.34.93
cryptography==42.0.7
PyJWT==2.8.0
gunicorn==22.0.0